Only show and save data for the user that is currently signed in

+++ TFG/DataBase/DBHelper.Sintrom.cs
@@ -8,6 +8,7 @@ namespace TFG.DataBase {
 		//Sintrom Health Module
         public static readonly string SINTROM_TABLE = "SINTROM";
         public static readonly string INR_TABLE = "INR_VALUES";
+        public static readonly string COL_USERID = "UserId";
         public static readonly string COL_IMAGENAME = "ImageName";
         public static readonly string COL_FRACTION = "Fraction";
         public static readonly string COL_MEDICINE = "Medicine";
@@ -17,15 +18,15 @@ namespace TFG.DataBase {
 		//Sintrom Health Module Queries
 
         public void CreateSintromTable() {
-            var sql = "CREATE TABLE IF NOT EXISTS " + SINTROM_TABLE + " (" + COL_DATE + " date primary key, "
-                + COL_IMAGENAME + " text, " + COL_FRACTION + " text, " + COL_MEDICINE + " text)";
+            var sql = "CREATE TABLE IF NOT EXISTS " + SINTROM_TABLE + " (" + COL_USERID + " text, " + COL_DATE + " date, "
+                + COL_IMAGENAME + " text, " + COL_FRACTION + " text, " + COL_MEDICINE + " text, primary key (" + COL_USERID + ", " + COL_DATE + "))";
 
             Connection.Execute(sql);
         }
 
         public void CreateINRTable() {
-            var sql = "CREATE TABLE IF NOT EXISTS " + INR_TABLE + " (" + COL_DATE + " date primary key, "
-                + COL_CONTROL + " boolean, " + COL_INR + " integer)";
+            var sql = "CREATE TABLE IF NOT EXISTS " + INR_TABLE + " (" + COL_USERID + " text, " + COL_DATE + " date, "
+                + COL_CONTROL + " boolean, " + COL_INR + " integer, primary key (" + COL_USERID + ", " + COL_DATE + "))";
 
             Connection.Execute(sql);
         }
@@ -34,16 +35,16 @@ namespace TFG.DataBase {
 
             var stringDate = sintromItem.Date.ToString(DATE_FORMAT);
 
-            var sql = "INSERT OR REPLACE INTO " + SINTROM_TABLE + " (" + COL_DATE + ", " + COL_IMAGENAME + ", " + COL_FRACTION + ", " + COL_MEDICINE + ") VALUES "
-                + "('" + stringDate + "', '" + sintromItem.ImageName + "', '" + sintromItem.Fraction + "', '" + sintromItem.Medicine + "')";
+            var sql = "INSERT OR REPLACE INTO " + SINTROM_TABLE + " (" + COL_USERID + ", " + COL_DATE + ", " + COL_IMAGENAME + ", " + COL_FRACTION + ", " + COL_MEDICINE + ") VALUES "
+                + "('" + sintromItem.UserId + "', '" + stringDate + "', '" + sintromItem.ImageName + "', '" + sintromItem.Fraction + "', '" + sintromItem.Medicine + "')";
 
             Connection.Execute(sql);
         }
 
         public void UpdateSintromItem(SintromTreatmentItem sintromItem) {
             if (sintromItem.ImageName.Equals("")) {
-                if (SintromItemExistsForDate(sintromItem.Date)) {
-                    RemoveSintromItemFromDate(sintromItem.Date);
+                if (SintromItemExistsForDate(sintromItem.Date, sintromItem.UserId)) {
+                    RemoveSintromItemFromDate(sintromItem.Date, sintromItem.UserId);
                 }
             } else {
                 InsertSintromItem(sintromItem);
@@ -51,33 +52,33 @@ namespace TFG.DataBase {
             
         }
 
-        public bool SintromItemExistsForDate(DateTime date) {
-            return GetSintromItemFromDate(date).Count > 0;
+        public bool SintromItemExistsForDate(DateTime date, string userId) {
+            return GetSintromItemFromDate(date, userId).Count > 0;
 
         }
 
         //Get Sintrom Treatment Item from a specific date
-        public List<SintromTreatmentItem> GetSintromItemFromDate(DateTime date) {
+        public List<SintromTreatmentItem> GetSintromItemFromDate(DateTime date, string userId) {
             var stringDate = date.ToString(DATE_FORMAT);
 
-            var sql = "SELECT * FROM " + SINTROM_TABLE + " WHERE " + COL_DATE + " = '" + stringDate + "'";
+            var sql = "SELECT * FROM " + SINTROM_TABLE + " WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " = '" + stringDate + "'";
 
             return Connection.Query<SintromTreatmentItem>(sql);
         }
 
-        public void RemoveSintromItemFromDate(DateTime date) {
+        public void RemoveSintromItemFromDate(DateTime date, string userId) {
 			var stringDate = date.ToString(DATE_FORMAT);
 
-            var sql = "DELETE FROM " + SINTROM_TABLE + " WHERE " + COL_DATE + " = '" + stringDate + "'";
+            var sql = "DELETE FROM " + SINTROM_TABLE + " WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " = '" + stringDate + "'";
 
 			Connection.Execute(sql);
         }
 
         //Get Sintrom Treatment Item from this date onwards
-        public List<SintromTreatmentItem> GetSintromItemsStartingFromDate(DateTime date) {
+        public List<SintromTreatmentItem> GetSintromItemsStartingFromDate(DateTime date, string userId) {
             var stringDate = date.ToString(DATE_FORMAT);
 
-            var sql = "SELECT * FROM " + SINTROM_TABLE + " WHERE " + COL_DATE + " >= '" + stringDate + "'";
+            var sql = "SELECT * FROM " + SINTROM_TABLE + " WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " >= '" + stringDate + "'";
 
             return Connection.Query<SintromTreatmentItem>(sql);
         }
@@ -86,30 +87,30 @@ namespace TFG.DataBase {
         public void InsertSintromINRItem(SintromINRItem sintromInrItem) {
             var stringDate = sintromInrItem.Date.ToString(DATE_FORMAT);
 
-            var sql = "INSERT OR REPLACE INTO " + INR_TABLE + " (" + COL_DATE + ", " + COL_CONTROL + ", " + COL_INR + ") VALUES "
-                + "('" + stringDate + "', '" + (sintromInrItem.Control ? "1" : "0")  + "', " + sintromInrItem.INR + ")";
+            var sql = "INSERT OR REPLACE INTO " + INR_TABLE + " (" + COL_USERID + ", " +COL_DATE + ", " + COL_CONTROL + ", " + COL_INR + ") VALUES "
+                + "('" + sintromInrItem.UserId + "', '" + stringDate + "', '" + (sintromInrItem.Control ? "1" : "0")  + "', " + sintromInrItem.INR + ")";
 
             Connection.Execute(sql);
         }
 
-        public void RemoveINRItemFromDate(DateTime date) {
+        public void RemoveINRItemFromDate(DateTime date, string userId) {
 			var stringDate = date.ToString(DATE_FORMAT);
 
-            var sql = "DELETE FROM " + INR_TABLE + " WHERE " + COL_DATE + " = '" + stringDate + "'";
+            var sql = "DELETE FROM " + INR_TABLE + " WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " = '" + stringDate + "'";
 
 			Connection.Execute(sql);
         }
 
 
-        public void RemoveOrHideSintromINRItem(DateTime date) {
-            var items = GetSintromINRItemFromDate(date);
+        public void RemoveOrHideSintromINRItem(DateTime date, string userId) {
+            var items = GetSintromINRItemFromDate(date, userId);
             if (items != null && items.Count > 0) {
                 var item = items[0];
                 if (item.INR == 0) {
-                    RemoveINRItemFromDate(date);
+                    RemoveINRItemFromDate(date, userId);
                 } else {
 					var stringDate = date.ToString(DATE_FORMAT);
-                    var sql = "UPDATE " + INR_TABLE + " SET " + COL_CONTROL + "= '0' WHERE " + COL_DATE + " = '" + stringDate + "'";
+                    var sql = "UPDATE " + INR_TABLE + " SET " + COL_CONTROL + "= '0' WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " = '" + stringDate + "'";
 					Connection.Execute(sql);
                 }
                 
@@ -123,24 +124,24 @@ namespace TFG.DataBase {
             return Connection.Query<SintromINRItem>(sql); 
         }
 
-        public List<SintromINRItem> GetSintromINRItemFromDate(DateTime date) {
+        public List<SintromINRItem> GetSintromINRItemFromDate(DateTime date, string userId) {
             var stringDate = date.ToString(DATE_FORMAT);
 
-            var sql = "SELECT * FROM " + INR_TABLE + " WHERE " + COL_DATE + " = '" + stringDate + "'";
+            var sql = "SELECT * FROM " + INR_TABLE + " WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " = '" + stringDate + "'";
 
             return Connection.Query<SintromINRItem>(sql);
         }
 
 		//Get Sintrom INR Item from this date onwards
-        public List<SintromINRItem> GetSintromINRItemsStartingFromDate(DateTime date) {
+        public List<SintromINRItem> GetSintromINRItemsStartingFromDate(DateTime date, string userId) {
             var stringDate = date.ToString(DATE_FORMAT);
 
-            var sql = "SELECT * FROM " + INR_TABLE + " WHERE " + COL_DATE + " >= '" + stringDate + "'";
+            var sql = "SELECT * FROM " + INR_TABLE + " WHERE " + COL_USERID + " = '" + userId + "' AND " + COL_DATE + " >= '" + stringDate + "'";
 
             return Connection.Query<SintromINRItem>(sql);
         }
 
-        private void FillSintromTable() {
+        private void FillSintromTable(string userId) {
             DropTable(SINTROM_TABLE);
             DropTable(INR_TABLE);
             CreateSintromTable();
@@ -154,7 +155,7 @@ namespace TFG.DataBase {
             for (var day = startDate.Date; day.Date <= endDate.Date; day = day.AddDays(1))  {
                 if (rnd.Next(10) == 1) {
                     //Control day
-                    InsertSintromINRItem(new SintromINRItem(day, true, 1.4));
+                    InsertSintromINRItem(new SintromINRItem(userId, day, true, 1.4));
                 } else {
                     var medicine = "Sintrom ";
                     var medicineRnd = rnd.Next(3);
@@ -174,7 +175,7 @@ namespace TFG.DataBase {
                         case 4: imageName += "1_8"; break;
                     }
 
-                    InsertSintromItem(new SintromTreatmentItem(day, imageName, medicine));
+                    InsertSintromItem(new SintromTreatmentItem(userId, day, imageName, medicine));
                 }
             }
